+++ src/main/java/io/myzticbean/finditemaddon/listeners/PWPlayerWarpRemoveEventListener.java
@@ -19,9 +19,8 @@
 package io.myzticbean.finditemaddon.listeners;
 
 import com.olziedev.playerwarps.api.events.warp.PlayerWarpRemoveEvent;
-import io.myzticbean.finditemaddon.dependencies.PlayerWarpsPlugin;
 import io.myzticbean.finditemaddon.FindItemAddOn;
-import org.bukkit.Bukkit;
+import io.myzticbean.finditemaddon.dependencies.PlayerWarpsPlugin;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
 
@@ -32,7 +31,7 @@ public class PWPlayerWarpRemoveEventListener implements Listener {
     @EventHandler
     public void onPlayerWarpRemove(PlayerWarpRemoveEvent event) {
         // Issue #24 Fix: Converted updateWarpsOnEventCall() call to async
-        Bukkit.getScheduler().runTaskAsynchronously(FindItemAddOn.getInstance(),
-                () -> PlayerWarpsPlugin.updateWarpsOnEventCall(event.getPlayerWarp(), true));
+        FindItemAddOn.getScheduler().runAsync(
+                (t) -> PlayerWarpsPlugin.updateWarpsOnEventCall(event.getPlayerWarp(), true));
     }
 }