+++ src/main/java/io/myzticbean/finditemaddon/quickshop/impl/QSReremakeAPIHandler.java
@@ -18,13 +18,13 @@
  */
 package io.myzticbean.finditemaddon.quickshop.impl;
 
-import io.myzticbean.finditemaddon.quickshop.QSApi;
-import io.myzticbean.finditemaddon.commands.quickshop.subcommands.FindItemCmdReremakeImpl;
 import io.myzticbean.finditemaddon.FindItemAddOn;
+import io.myzticbean.finditemaddon.commands.quickshop.subcommands.FindItemCmdReremakeImpl;
 import io.myzticbean.finditemaddon.models.CachedShop;
 import io.myzticbean.finditemaddon.models.FoundShopItemModel;
 import io.myzticbean.finditemaddon.models.ShopSearchActivityModel;
 import io.myzticbean.finditemaddon.models.enums.PlayerPermsEnum;
+import io.myzticbean.finditemaddon.quickshop.QSApi;
 import io.myzticbean.finditemaddon.utils.json.HiddenShopStorageUtil;
 import io.myzticbean.finditemaddon.utils.log.Logger;
 import org.bukkit.Bukkit;
@@ -46,6 +46,7 @@ import java.util.Date;
 import java.util.List;
 import java.util.Objects;
 import java.util.UUID;
+import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ConcurrentMap;
 
@@ -70,124 +71,117 @@ public class QSReremakeAPIHandler implements QSApi<QuickShop, Shop> {
     }
 
     @Override
-    public List<FoundShopItemModel> findItemBasedOnTypeFromAllShops(ItemStack item, boolean toBuy, Player searchingPlayer) {
+    public CompletableFuture<List<FoundShopItemModel>> findItemBasedOnTypeFromAllShops(ItemStack item, boolean toBuy, Player searchingPlayer) {
         var begin = Instant.now();
         List<FoundShopItemModel> shopsFoundList = new ArrayList<>();
-        List<Shop> allShops;
-        if(FindItemAddOn.getConfigProvider().SEARCH_LOADED_SHOPS_ONLY) {
-            allShops = new ArrayList<>(api.getShopManager().getLoadedShops());
-        }
-        else {
-            allShops = api.getShopManager().getAllShops();
-        }
-        logTotalShopsOnServer(allShops.size());
-        for(Shop shop_i : allShops) {
-            // check for blacklisted worlds
-            if(!FindItemAddOn.getConfigProvider().getBlacklistedWorlds().contains(shop_i.getLocation().getWorld())
-                    && shop_i.getItem().getType().equals(item.getType())
-                    && (toBuy ? shop_i.isSelling() : shop_i.isBuying())) {
-                if(checkIfShopToBeIgnoredForFullOrEmpty(toBuy, shop_i)) {
-                    continue;
-                }
-                // check for shop if hidden
-                if(!HiddenShopStorageUtil.isShopHidden(shop_i)) {
-                    shopsFoundList.add(new FoundShopItemModel(
-                            shop_i.getPrice(),
-                            QSApi.processStockOrSpace((toBuy ? getRemainingStockOrSpaceFromShopCache(shop_i, true) : getRemainingStockOrSpaceFromShopCache(shop_i, false))),
-                            shop_i.getOwner(),
-                            shop_i.getLocation(),
-                            shop_i.getItem(),
-                            toBuy
-                    ));
-                }
-            }
-        }
+//        List<Shop> allShops;
+//        if(FindItemAddOn.getConfigProvider().SEARCH_LOADED_SHOPS_ONLY) {
+//            allShops = new ArrayList<>(api.getShopManager().getLoadedShops());
+//        } else {
+//            allShops = api.getShopManager().getAllShops();
+//        }
+//        logTotalShopsOnServer(allShops.size());
+//        for(Shop shop : allShops) {
+//            // check for blacklisted worlds
+//            if(!FindItemAddOn.getConfigProvider().getBlacklistedWorlds().contains(shop.getLocation().getWorld())
+//                    && shop.getItem().getType().equals(item.getType())
+//                    && (toBuy ? shop.isSelling() : shop.isBuying())) {
+//                if(checkIfShopToBeIgnoredForFullOrEmpty(toBuy, shop)) {
+//                    continue;
+//                }
+//                // check for shop if hidden
+//                if(!HiddenShopStorageUtil.isShopHidden(shop)) {
+//                    shopsFoundList.add(new FoundShopItemModel(
+//                            shop.getPrice(),
+//                            QSApi.processStockOrSpace((toBuy ? getRemainingStockOrSpaceFromShopCache(shop, true) : getRemainingStockOrSpaceFromShopCache(shop, false))),
+//                            shop.getOwner(),
+//                            shop.getLocation(),
+//                            shop.getItem(),
+//                            toBuy
+//                    ));
+//                }
+//            }
+//        }
         List<FoundShopItemModel> sortedShops = handleShopSorting(toBuy, shopsFoundList);
         QSApi.logTimeTookMsg(begin);
-        return sortedShops;
+        return CompletableFuture.completedFuture(sortedShops);
     }
 
     @Override
-    public List<FoundShopItemModel> findItemBasedOnDisplayNameFromAllShops(String displayName, boolean toBuy, Player searchingPlayer) {
+    public CompletableFuture<List<FoundShopItemModel>> findItemBasedOnDisplayNameFromAllShops(String displayName, boolean toBuy, Player searchingPlayer) {
         var begin = Instant.now();
         List<FoundShopItemModel> shopsFoundList = new ArrayList<>();
-        List<Shop> allShops;
-        if(FindItemAddOn.getConfigProvider().SEARCH_LOADED_SHOPS_ONLY) {
-            allShops = new ArrayList<>(api.getShopManager().getLoadedShops());
-        }
-        else {
-            allShops = api.getShopManager().getAllShops();
-        }
-        logTotalShopsOnServer(allShops.size());
-        for(Shop shop_i : allShops) {
-            if(!FindItemAddOn.getConfigProvider().getBlacklistedWorlds().contains(shop_i.getLocation().getWorld())) {
-                if(shop_i.getItem().hasItemMeta()) {
-                    if(Objects.requireNonNull(shop_i.getItem().getItemMeta()).hasDisplayName()) {
-                        if(shop_i.getItem().getItemMeta().getDisplayName().toLowerCase().contains(displayName.toLowerCase())
-//                                && (toBuy ? getRemainingStockOrSpaceFromShopCache(shop_i, true) != 0 : getRemainingStockOrSpaceFromShopCache(shop_i, false) != 0)
-                                && (toBuy ? shop_i.isSelling() : shop_i.isBuying())) {
-                            if(checkIfShopToBeIgnoredForFullOrEmpty(toBuy, shop_i))
-                                continue;
-                            // check for shop if hidden
-                            if(!HiddenShopStorageUtil.isShopHidden(shop_i)) {
-                                shopsFoundList.add(new FoundShopItemModel(
-                                        shop_i.getPrice(),
-                                        QSApi.processStockOrSpace((toBuy ? getRemainingStockOrSpaceFromShopCache(shop_i, true) : getRemainingStockOrSpaceFromShopCache(shop_i, false))),
-                                        shop_i.getOwner(),
-                                        shop_i.getLocation(),
-                                        shop_i.getItem(),
-                                        toBuy
-                                ));
-                            }
-                        }
-                    }
-                }
-            }
-        }
+//        List<Shop> allShops;
+//        if(FindItemAddOn.getConfigProvider().SEARCH_LOADED_SHOPS_ONLY) {
+//            allShops = new ArrayList<>(api.getShopManager().getLoadedShops());
+//        } else {
+//            allShops = api.getShopManager().getAllShops();
+//        }
+//        logTotalShopsOnServer(allShops.size());
+//        for(Shop shop : allShops) {
+//            if(!FindItemAddOn.getConfigProvider().getBlacklistedWorlds().contains(shop.getLocation().getWorld())
+//                    && shop.getItem().hasItemMeta()
+//                    && Objects.requireNonNull(shop.getItem().getItemMeta()).hasDisplayName()
+//                    && shop.getItem().getItemMeta().getDisplayName().toLowerCase().contains(displayName.toLowerCase())
+////                  && (toBuy ? getRemainingStockOrSpaceFromShopCache(shop, true) != 0 : getRemainingStockOrSpaceFromShopCache(shop, false) != 0)
+//                    && (toBuy ? shop.isSelling() : shop.isBuying())) {
+//                if(checkIfShopToBeIgnoredForFullOrEmpty(toBuy, shop)) continue;
+//                // check for shop if hidden
+//                if(!HiddenShopStorageUtil.isShopHidden(shop)) {
+//                    shopsFoundList.add(new FoundShopItemModel(
+//                            shop.getPrice(),
+//                            QSApi.processStockOrSpace((toBuy ? getRemainingStockOrSpaceFromShopCache(shop, true) : getRemainingStockOrSpaceFromShopCache(shop, false))),
+//                            shop.getOwner(),
+//                            shop.getLocation(),
+//                            shop.getItem(),
+//                            toBuy
+//                    ));
+//                }
+//            }
+//        }
         List<FoundShopItemModel> sortedShops = handleShopSorting(toBuy, shopsFoundList);
         QSApi.logTimeTookMsg(begin);
-        return sortedShops;
+        return CompletableFuture.completedFuture(sortedShops);
     }
 
     @Override
-    public List<FoundShopItemModel> fetchAllItemsFromAllShops(boolean toBuy, Player searchingPlayer) {
+    public CompletableFuture<List<FoundShopItemModel>> fetchAllItemsFromAllShops(boolean toBuy, Player searchingPlayer) {
         var begin = Instant.now();
         List<FoundShopItemModel> shopsFoundList = new ArrayList<>();
-        List<Shop> allShops;
-        if(FindItemAddOn.getConfigProvider().SEARCH_LOADED_SHOPS_ONLY) {
-            allShops = new ArrayList<>(api.getShopManager().getLoadedShops());
-        }
-        else {
-            allShops = api.getShopManager().getAllShops();
-        }
-        logTotalShopsOnServer(allShops.size());
-        for(Shop shop_i : allShops) {
-            // check for blacklisted worlds
-            if(!FindItemAddOn.getConfigProvider().getBlacklistedWorlds().contains(shop_i.getLocation().getWorld())
-//                    && (toBuy ? getRemainingStockOrSpaceFromShopCache(shop_i, true) != 0 : getRemainingStockOrSpaceFromShopCache(shop_i, false) != 0)
-                    && (toBuy ? shop_i.isSelling() : shop_i.isBuying())) {
-                if(checkIfShopToBeIgnoredForFullOrEmpty(toBuy, shop_i))
-                    continue;
-                // check for shop if hidden
-                if(!HiddenShopStorageUtil.isShopHidden(shop_i)) {
-                    shopsFoundList.add(new FoundShopItemModel(
-                            shop_i.getPrice(),
-                            QSApi.processStockOrSpace((toBuy ? getRemainingStockOrSpaceFromShopCache(shop_i, true) : getRemainingStockOrSpaceFromShopCache(shop_i, false))),
-                            shop_i.getOwner(),
-                            shop_i.getLocation(),
-                            shop_i.getItem(),
-                            toBuy
-                    ));
-                }
-            }
-        }
+//        List<Shop> allShops;
+//        if(FindItemAddOn.getConfigProvider().SEARCH_LOADED_SHOPS_ONLY) {
+//            allShops = new ArrayList<>(api.getShopManager().getLoadedShops());
+//        } else {
+//            allShops = api.getShopManager().getAllShops();
+//        }
+//        logTotalShopsOnServer(allShops.size());
+//        for(Shop shop : allShops) {
+//            // check for blacklisted worlds
+//            if(!FindItemAddOn.getConfigProvider().getBlacklistedWorlds().contains(shop.getLocation().getWorld())
+////                  && (toBuy ? getRemainingStockOrSpaceFromShopCache(shop, true) != 0 : getRemainingStockOrSpaceFromShopCache(shop, false) != 0)
+//                    && (toBuy ? shop.isSelling() : shop.isBuying())) {
+//                if(checkIfShopToBeIgnoredForFullOrEmpty(toBuy, shop))
+//                    continue;
+//                // check for shop if hidden
+//                if(!HiddenShopStorageUtil.isShopHidden(shop)) {
+//                    shopsFoundList.add(new FoundShopItemModel(
+//                            shop.getPrice(),
+//                            QSApi.processStockOrSpace((toBuy ? getRemainingStockOrSpaceFromShopCache(shop, true) : getRemainingStockOrSpaceFromShopCache(shop, false))),
+//                            shop.getOwner(),
+//                            shop.getLocation(),
+//                            shop.getItem(),
+//                            toBuy
+//                    ));
+//                }
+//            }
+//        }
         List<FoundShopItemModel> sortedShops = new ArrayList<>(shopsFoundList);
-        if(!shopsFoundList.isEmpty()) {
-            int sortingMethod = 1;
-            sortedShops = QSApi.sortShops(sortingMethod, shopsFoundList, toBuy);
-        }
+//        if(!shopsFoundList.isEmpty()) {
+//            int sortingMethod = 1;
+//            sortedShops = QSApi.sortShops(sortingMethod, shopsFoundList, toBuy);
+//        }
         QSApi.logTimeTookMsg(begin);
-        return sortedShops;
+        return CompletableFuture.completedFuture(sortedShops);
     }
 
     @Override
@@ -217,39 +211,39 @@ public class QSReremakeAPIHandler implements QSApi<QuickShop, Shop> {
         // copy all shops from shops list in API to a temp globalShopsList
         // now check shops from temp globalShopsList in current globalShopsList and pull playerVisit data
         List<ShopSearchActivityModel> tempGlobalShopsList = new ArrayList<>();
-
-        for(Shop shop_i : getAllShops()) {
-            Location shopLoc = shop_i.getLocation();
-            tempGlobalShopsList.add(new ShopSearchActivityModel(
-                    shopLoc.getWorld().getName(),
-                    shopLoc.getX(),
-                    shopLoc.getY(),
-                    shopLoc.getZ(),
-                    shopLoc.getPitch(),
-                    shopLoc.getYaw(),
-                    shop_i.getOwner().toString(),
-                    new ArrayList<>(),
-                    false
-            ));
-        }
-        for(ShopSearchActivityModel shop_temp : tempGlobalShopsList) {
-            ShopSearchActivityModel tempShopToRemove = null;
-            for(ShopSearchActivityModel shop_global : globalShopsList) {
-                if(shop_temp.getWorldName().equalsIgnoreCase(shop_global.getWorldName())
-                        && shop_temp.getX() == shop_global.getX()
-                        && shop_temp.getY() == shop_global.getY()
-                        && shop_temp.getZ() == shop_global.getZ()
-                        && shop_temp.getShopOwnerUUID().equalsIgnoreCase(shop_global.getShopOwnerUUID())
-                ) {
-                    shop_temp.setPlayerVisitList(shop_global.getPlayerVisitList());
-                    shop_temp.setHiddenFromSearch(shop_global.isHiddenFromSearch());
-                    tempShopToRemove = shop_global;
-                    break;
-                }
-            }
-            if(tempShopToRemove != null)
-                globalShopsList.remove(tempShopToRemove);
-        }
+//
+//        for(Shop shop_i : getAllShops()) {
+//            Location shopLoc = shop_i.getLocation();
+//            tempGlobalShopsList.add(new ShopSearchActivityModel(
+//                    shopLoc.getWorld().getName(),
+//                    shopLoc.getX(),
+//                    shopLoc.getY(),
+//                    shopLoc.getZ(),
+//                    shopLoc.getPitch(),
+//                    shopLoc.getYaw(),
+//                    shop_i.getOwner().toString(),
+//                    new ArrayList<>(),
+//                    false
+//            ));
+//        }
+//        for(ShopSearchActivityModel shop_temp : tempGlobalShopsList) {
+//            ShopSearchActivityModel tempShopToRemove = null;
+//            for(ShopSearchActivityModel shop_global : globalShopsList) {
+//                if(shop_temp.getWorldName().equalsIgnoreCase(shop_global.getWorldName())
+//                        && shop_temp.getX() == shop_global.getX()
+//                        && shop_temp.getY() == shop_global.getY()
+//                        && shop_temp.getZ() == shop_global.getZ()
+//                        && shop_temp.getShopOwnerUUID().equalsIgnoreCase(shop_global.getShopOwnerUUID())
+//                ) {
+//                    shop_temp.setPlayerVisitList(shop_global.getPlayerVisitList());
+//                    shop_temp.setHiddenFromSearch(shop_global.isHiddenFromSearch());
+//                    tempShopToRemove = shop_global;
+//                    break;
+//                }
+//            }
+//            if(tempShopToRemove != null)
+//                globalShopsList.remove(tempShopToRemove);
+//        }
         return tempGlobalShopsList;
     }
 
@@ -258,22 +252,22 @@ public class QSReremakeAPIHandler implements QSApi<QuickShop, Shop> {
      */
     @Override
     public void registerSubCommand() {
-        Logger.logInfo("Unregistered find sub-command for /qs");
-        for(CommandContainer cmdContainer : api.getCommandManager().getRegisteredCommands()) {
-            if(cmdContainer.getPrefix().equalsIgnoreCase("find")) {
-                api.getCommandManager().unregisterCmd(cmdContainer);
-                break;
-            }
-        }
-        Logger.logInfo("Registered finditem sub-command for /qs");
-        api.getCommandManager().registerCmd(
-                CommandContainer.builder()
-                        .prefix("finditem")
-                        .permission(PlayerPermsEnum.FINDITEM_USE.value())
-                        .hidden(false)
-                        .description("Search for items from all shops using an interactive GUI")
-                        .executor(new FindItemCmdReremakeImpl())
-                        .build());
+//        Logger.logInfo("Unregistered find sub-command for /qs");
+//        for(CommandContainer cmdContainer : api.getCommandManager().getRegisteredCommands()) {
+//            if(cmdContainer.getPrefix().equalsIgnoreCase("find")) {
+//                api.getCommandManager().unregisterCmd(cmdContainer);
+//                break;
+//            }
+//        }
+//        Logger.logInfo("Registered finditem sub-command for /qs");
+//        api.getCommandManager().registerCmd(
+//                CommandContainer.builder()
+//                        .prefix("finditem")
+//                        .permission(PlayerPermsEnum.FINDITEM_USE.value())
+//                        .hidden(false)
+//                        .description("Search for items from all shops using an interactive GUI")
+//                        .executor(new FindItemCmdReremakeImpl())
+//                        .build());
     }
 
     @Override
@@ -283,17 +277,17 @@ public class QSReremakeAPIHandler implements QSApi<QuickShop, Shop> {
 
     @NotNull
     static List<FoundShopItemModel> handleShopSorting(boolean toBuy, List<FoundShopItemModel> shopsFoundList) {
-        if(!shopsFoundList.isEmpty()) {
-            int sortingMethod = 2;
-            try {
-                sortingMethod = FindItemAddOn.getConfigProvider().SHOP_SORTING_METHOD;
-            }
-            catch(Exception e) {
-                Logger.logError("Invalid value in config.yml : 'shop-sorting-method'");
-                Logger.logError("Defaulting to sorting by prices method");
-            }
-            return QSApi.sortShops(sortingMethod, shopsFoundList, toBuy);
-        }
+//        if(!shopsFoundList.isEmpty()) {
+//            int sortingMethod = 2;
+//            try {
+//                sortingMethod = FindItemAddOn.getConfigProvider().SHOP_SORTING_METHOD;
+//            }
+//            catch(Exception e) {
+//                Logger.logError("Invalid value in config.yml : 'shop-sorting-method'");
+//                Logger.logError("Defaulting to sorting by prices method");
+//            }
+//            return QSApi.sortShops(sortingMethod, shopsFoundList, toBuy);
+//        }
         return shopsFoundList;
     }
 
@@ -306,14 +300,14 @@ public class QSReremakeAPIHandler implements QSApi<QuickShop, Shop> {
      * @return If shop needs to be ignored from list
      */
     private boolean checkIfShopToBeIgnoredForFullOrEmpty(boolean toBuy, Shop shop) {
-        boolean ignoreEmptyChests = FindItemAddOn.getConfigProvider().IGNORE_EMPTY_CHESTS;
-        if(ignoreEmptyChests) {
-            if(toBuy) {
-                return getRemainingStockOrSpaceFromShopCache(shop, true) == 0;
-
-            } else return getRemainingStockOrSpaceFromShopCache(shop, false) == 0;
-
-        }
+//        boolean ignoreEmptyChests = FindItemAddOn.getConfigProvider().IGNORE_EMPTY_CHESTS;
+//        if(ignoreEmptyChests) {
+//            if(toBuy) {
+//                return getRemainingStockOrSpaceFromShopCache(shop, true) == 0;
+//
+//            } else return getRemainingStockOrSpaceFromShopCache(shop, false) == 0;
+//
+//        }
         return false;
     }
 
@@ -338,29 +332,30 @@ public class QSReremakeAPIHandler implements QSApi<QuickShop, Shop> {
      * @return
      */
     private int getRemainingStockOrSpaceFromShopCache(Shop shop, boolean fetchRemainingStock) {
-        Logger.logDebugInfo("Shop Location: " + shop.getLocation());
-        CachedShop cachedShop = shopCache.get(shop.getLocation());
-        if (cachedShop == null || QSApi.isTimeDifferenceGreaterThanSeconds(cachedShop.getLastFetched(), new Date(), SHOP_CACHE_TIMEOUT_SECONDS)) {
-            Shop possibleCachedShop = getRemainingStockOrSpaceFromQSCache(shop);
-            if(possibleCachedShop != null) {
-                Logger.logDebugInfo("Shop found from QS cache: " + shop.getLocation());
-                cachedShop = CachedShop.builder()
-                                .shopLocation(possibleCachedShop.getLocation())
-                                .remainingStock(possibleCachedShop.getRemainingStock())
-                                .remainingSpace(possibleCachedShop.getRemainingSpace())
-                                .lastFetched(new Date())
-                                .build();
-                shopCache.put(cachedShop.getShopLocation(), cachedShop);
-                Logger.logDebugInfo("Added to ShopCache: " + shop.getLocation());
-            } else {
-                Logger.logDebugInfo("No shop found from QS cache for location: " + shop.getLocation());
-            }
-        } else {
-            Logger.logDebugInfo("Shop found from cache: " + shop.getLocation());
-        }
-        return (fetchRemainingStock ?
-                (cachedShop != null ? cachedShop.getRemainingStock() : -1)
-                : (cachedShop != null ? cachedShop.getRemainingSpace() : -1));
+//        Logger.logDebugInfo("Shop Location: " + shop.getLocation());
+//        CachedShop cachedShop = shopCache.get(shop.getLocation());
+//        if (cachedShop == null || QSApi.isTimeDifferenceGreaterThanSeconds(cachedShop.getLastFetched(), new Date(), SHOP_CACHE_TIMEOUT_SECONDS)) {
+//            Shop possibleCachedShop = getRemainingStockOrSpaceFromQSCache(shop);
+//            if(possibleCachedShop != null) {
+//                Logger.logDebugInfo("Shop found from QS cache: " + shop.getLocation());
+//                cachedShop = CachedShop.builder()
+//                                .shopLocation(possibleCachedShop.getLocation())
+//                                .remainingStock(possibleCachedShop.getRemainingStock())
+//                                .remainingSpace(possibleCachedShop.getRemainingSpace())
+//                                .lastFetched(new Date())
+//                                .build();
+//                shopCache.put(cachedShop.getShopLocation(), cachedShop);
+//                Logger.logDebugInfo("Added to ShopCache: " + shop.getLocation());
+//            } else {
+//                Logger.logDebugInfo("No shop found from QS cache for location: " + shop.getLocation());
+//            }
+//        } else {
+//            Logger.logDebugInfo("Shop found from cache: " + shop.getLocation());
+//        }
+//        return (fetchRemainingStock ?
+//                (cachedShop != null ? cachedShop.getRemainingStock() : -1)
+//                : (cachedShop != null ? cachedShop.getRemainingSpace() : -1));
+        return -1;
     }
 
     /**